requirements.txt
google-adk
yfinance==0.2.56
psutil==5.9.5
litellm==1.66.3
google-generativeai
python-dotenv==1.1.0
deprecated
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
streamlit>=1.28.0
uuid
